fix(SBM): Aggiornata struttura progetto
Aggiornata struttura del progetto per adeguarsi alle nuove modifiche.
Il codice dei Probabilistic Methods ora è all'interno di un singolo
file.

Aggiornata documentazione
