--- cnlp/probabilistic_methods/__init__.py
@@ -1,1 +1,0 @@
-from .sbm import stochastic_block_model
