+++ docs/index.html
@@ -35,9 +35,9 @@
 <dd>
 <div class="desc"></div>
 </dd>
-<dt><code class="name"><a title="cnlp.probabilistic_methods" href="probabilistic_methods/index.html">cnlp.probabilistic_methods</a></code></dt>
+<dt><code class="name"><a title="cnlp.probabilistic_methods" href="probabilistic_methods.html">cnlp.probabilistic_methods</a></code></dt>
 <dd>
-<div class="desc"></div>
+<div class="desc"><p>Collection of Probabilistic Methods for Link Prediction.</p></div>
 </dd>
 <dt><code class="name"><a title="cnlp.similarity_methods" href="similarity_methods/index.html">cnlp.similarity_methods</a></code></dt>
 <dd>
@@ -66,7 +66,7 @@
 <ul>
 <li><code><a title="cnlp.dimensionality_reduction_methods" href="dimensionality_reduction_methods.html">cnlp.dimensionality_reduction_methods</a></code></li>
 <li><code><a title="cnlp.other_methods" href="other_methods/index.html">cnlp.other_methods</a></code></li>
-<li><code><a title="cnlp.probabilistic_methods" href="probabilistic_methods/index.html">cnlp.probabilistic_methods</a></code></li>
+<li><code><a title="cnlp.probabilistic_methods" href="probabilistic_methods.html">cnlp.probabilistic_methods</a></code></li>
 <li><code><a title="cnlp.similarity_methods" href="similarity_methods/index.html">cnlp.similarity_methods</a></code></li>
 <li><code><a title="cnlp.utils" href="utils.html">cnlp.utils</a></code></li>
 </ul>
